CDEC Sensor Details (by Station)

Query CDEC Website for Sensor Details

CDEC_StationInfo(s)

Arguments

s
a CDEC station ID (e.g. 'HHM')

Details

This function requires the `rvest` package

Value

a `list` object containing site metadata, sensor metadata, and possibly comments about the site

Author

D.E. Beaudette

See also

Examples

# \donttest{
if(requireNamespace("curl") &
   curl::has_internet() 
   ) {
  
CDEC_StationInfo('HHM')  
  
}

#> $site.meta
#>   Station ID River Basin   Hydrologic Area Latitude
#> 1        HHM MOKELUMNE R SAN JOAQUIN RIVER 38.49066
#>                                   Operator Elevation County Nearby City
#> 1 CA Dept of Water Resources/DFM-Hydro-SMN      8700 ALPINE BEAR VALLEY
#>   Longitude Data Collection            Name
#> 1  119.8056                 HIGHLAND MEADOW
#> 
#> $sensor.meta
#>                          sensor_details sensor  interval sensor_name
#> 1                    SNOW DEPTH, INCHES     18 (monthly)   (SNOW DP)
#> 2           SNOW, WATER CONTENT, INCHES      3 (monthly)   (SNOW WC)
#> 3  SNOW, WATER CONTENT(REVISED), INCHES     82 (monthly)   (SNO ADJ)
#> 4    PRECIPITATION, ACCUMULATED, INCHES      2   (daily)      (RAIN)
#> 5    PRECIPITATION, INCREMENTAL, INCHES     45   (daily)   (PPT INC)
#> 6           SNOW, WATER CONTENT, INCHES      3   (daily)   (SNOW WC)
#> 7  SNOW, WATER CONTENT(REVISED), INCHES     82   (daily)   (SNO ADJ)
#> 8       TEMPERATURE, AIR AVERAGE, DEG F     30   (daily)   (TEMP AV)
#> 9       TEMPERATURE, AIR MAXIMUM, DEG F     31   (daily)   (TEMP MX)
#> 10      TEMPERATURE, AIR MINIMUM, DEG F     32   (daily)   (TEMP MN)
#> 11         ATMOSPHERIC PRESSURE, INCHES     17   (event)   (BAR PRE)
#> 12               BATTERY VOLTAGE, VOLTS     14   (event)   (BAT VOL)
#> 13   PRECIPITATION, ACCUMULATED, INCHES      2   (event)      (RAIN)
#> 14      REFLECTED IRRADIANCE AVG, W/M^2    119   (event)   (RIRR AV)
#> 15               RELATIVE HUMIDITY,   %     12   (event)   (REL HUM)
#> 16                   SNOW DEPTH, INCHES     18   (event)   (SNOW DP)
#> 17          SNOW, WATER CONTENT, INCHES      3   (event)   (SNOW WC)
#> 18              SOIL MOISTR, DEPTH 1, %    197   (event)   (SOILMD1)
#> 19              SOIL MOISTR, DEPTH 2, %    198   (event)   (SOILMD2)
#> 20              SOIL MOISTR, DEPTH 3, %    199   (event)   (SOILMD3)
#> 21            SOIL TEMP, DEPTH 1, DEG F    194   (event)   (SOILTD1)
#> 22            SOIL TEMP, DEPTH 2, DEG F    195   (event)   (SOILTD2)
#> 23            SOIL TEMP, DEPTH 3, DEG F    196   (event)   (SOILTD3)
#> 24               SOLAR RADIATION, W/M^2     26   (event)   (SOLAR R)
#> 25              TEMPERATURE, AIR, DEG F      4   (event)      (TEMP)
#> 26                 WIND, DIRECTION, DEG     10   (event)   (WIND DR)
#> 27                     WIND, SPEED, MPH      9   (event)   (WIND SP)
#> 28               BATTERY VOLTAGE, VOLTS     14  (hourly)   (BAT VOL)
#> 29   PRECIPITATION, ACCUMULATED, INCHES      2  (hourly)      (RAIN)
#> 30          SNOW, WATER CONTENT, INCHES      3  (hourly)   (SNOW WC)
#> 31              TEMPERATURE, AIR, DEG F      4  (hourly)      (TEMP)
#> 32           SOLAR RADIATION AVG, W/M^2    103  (hourly)   (SLRR AV)
#> 33           SOLAR RADIATION MAX, W/M^2    105  (hourly)   (SLRR MX)
#> 34           SOLAR RADIATION MIN, W/M^2    104  (hourly)   (SLRR MN)
#>    collection_method         period_of_record
#> 1       MANUAL ENTRY    05/01/1952 to present
#> 2       MANUAL ENTRY    05/01/1952 to present
#> 3       MANUAL ENTRY    05/01/1952 to present
#> 4           COMPUTED    10/01/2002 to present
#> 5           COMPUTED    12/01/1996 to present
#> 6           COMPUTED    11/29/1989 to present
#> 7           COMPUTED    10/21/1989 to present
#> 8           COMPUTED    09/02/1999 to present
#> 9           COMPUTED    09/02/1999 to present
#> 10          COMPUTED    09/02/1999 to present
#> 11         SATELLITE    09/14/2016 to present
#> 12         SATELLITE    09/14/2016 to present
#> 13         SATELLITE    09/14/2016 to present
#> 14         SATELLITE    09/14/2016 to present
#> 15         SATELLITE    09/14/2016 to present
#> 16         SATELLITE    09/14/2016 to present
#> 17         SATELLITE    09/14/2016 to present
#> 18         SATELLITE    09/14/2016 to present
#> 19         SATELLITE    09/14/2016 to present
#> 20         SATELLITE    09/14/2016 to present
#> 21         SATELLITE    09/14/2016 to present
#> 22         SATELLITE    09/14/2016 to present
#> 23         SATELLITE    09/14/2016 to present
#> 24         SATELLITE    09/14/2016 to present
#> 25         SATELLITE    09/14/2016 to present
#> 26         SATELLITE    09/14/2016 to present
#> 27         SATELLITE    09/14/2016 to present
#> 28         SATELLITE    01/01/1995 to present
#> 29         SATELLITE    10/14/1989 to present
#> 30         SATELLITE    10/14/1989 to present
#> 31         SATELLITE    10/13/1989 to present
#> 32         SATELLITE 10/10/2012 to 09/14/2016
#> 33         SATELLITE 10/10/2012 to 09/14/2016
#> 34         SATELLITE 10/10/2012 to 09/14/2016
#> 
#> $comments
#>         date
#> 1 04/27/2020
#> 2 04/27/2020
#> 3 07/23/2014
#>                                                                                                               comment
#> 1 Latitude and longitude updated from information provided by DFM-Statewide Monitoring Network Section.  Datum WGS-84
#> 2 Latitude and longitude updated from information provided by DFM-Statewide Monitoring Network Section.  Datum WGS-84
#> 3                        Latitude and longitude updated based on information from Hydrology Branch field technicians.
#> 

# }