LL2PLSS

Uses lattitude and longitude coordinates to return the PLSS section geometry from the BLM PLSS web service.

LL2PLSS(x, y, returnlevel = "I")

Arguments

x
longitude coordinates

y
lattitude coordinates

returnlevel
'S' for "Section" or 'I' for "Intersection" (subsections)

Value

list of of PLSS codes and coordinates.

Details

This function takes xy coordinates and returns the PLSS section geometry to the quarter-quarter section. returnlevel options are defaulted to 'I' which returns smallest intersected sectional aliquot geometry, 'S' will return the section geometry of the coordinates. See https://gis.blm.gov/arcgis/rest/services/Cadastral/BLM_Natl_PLSS_CadNSDI/MapServer for details.

Note

This function requires the following packages: httr, jsonlite, and sp.

See also

Author

D.E. Beaudette, Jay Skovlin, A.G. Brown