PLSS2LL

Fetch lattitude and longitude centroid coordinates for coded PLSS information from the BLM PLSS web service.

PLSS2LL(p)

Arguments

p
dataframe with chunks of PLSS coordinates

Details

This function takes coded PLSS information and fetches the centroid lat/long coordinates down to the quarter-quarter section. The 'plssid' column is generated within a dataframe using the following formatPLSS function.

Value

A dataframe of PLSS codes and coordinates.

Note

This function expects that the dataframe will have a 'plssid' column generated by the formatPLSS function. Requires the following packages: httr, and jsonlite.

See also

Examples

# \donttest{

if(requireNamespace("curl") &
   curl::has_internet()
   ) {


# create some data
d <- data.frame(
id=1:3,
qq=c('SW', 'SW', 'SE'),
q=c('NE', 'NW', 'SE'),
s=c(17, 32, 30),
t=c('T36N', 'T35N', 'T35N'),
r=c('R29W', 'R28W', 'R28W'),
type='SN',
m='MT20', stringsAsFactors = FALSE)

# add column names
names(d) <- c('id', 'qq', 'q', 's', 't', 'r', 'type', 'm')

# generate formatted PLSS codes
d$plssid <- formatPLSS(d)

# fetch lat/long coordinates
PLSS2LL(d)

}

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================================| 100%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================================| 100%  |                                                                              |=======================                                               |  33%  |                                                                              |===============================================                       |  67%  |                                                                              |======================================================================| 100%

#>   id                    plssid      lat       lon
#> 1  1 MT200360N0290W0SN170ASWNE 48.88590 -115.3768
#> 2  2 MT200350N0280W0SN320ASWNW 48.75450 -115.2578
#> 3  3 MT200350N0280W0SN300ASESE 48.76174 -115.2633


# }