component.adj.matrix {sharpshootR}	R Documentation

Create an adjacency matrix from a data.frame of component data

Description

Create an adjacency matrix from a data.frame of component data

Usage

component.adj.matrix(d, mu = "mukey", co = "compname", wt=NULL)

Arguments

d
a data.frame

mu
name of the column containing the map unit ID (typically 'mukey')

co
name of the column containing the component ID (typically 'compname')

wt
name of the column containing the component weight percent (typically 'comppct_r')

Value

an ajacency matrix suitable for use with igraph functions

Author(s)

D.E. Beaudette

Examples

# load sample data set
data(amador)

# convert into unweighted adjacency matrix
m <- component.adj.matrix(amador)

# plot network diagram, with Amador soil highlighted
plotSoilRelationGraph(m, s='amador')

## Error: could not find function "membership"

[Package sharpshootR version 0.3 Index]