Perform daily water balance modeling using SSURGO and DAYMET

Pending.

dailyWB_SSURGO(
  x,
  cokeys = NULL,
  start = 1988,
  end = 2018,
  modelDepth = 100,
  MS.style = "default",
  a.ss = 0.1,
  S_0 = 0.5,
  bufferRadiusMeters = 1
)

Arguments

x

SpatialPoints object representing a single point

cokeys

vector of component keys to use

start

starting year (limited to DAYMET holdings)

end

ending year (limited to DAYMET holdings)

modelDepth

soil depth used for water balance, see details

MS.style

moisture state classification style, see estimateSoilMoistureState

a.ss

recession coefficients for subsurface flow from saturated zone, should be > 0 (range: 0-1)

S_0

fraction of water storage filled at time = 0 (range: 0-1)

bufferRadiusMeters

spatial buffer (meters) applied to x for the look-up of SSURGO data

Value

data.frame of daily water balance results

References

Farmer, D., M. Sivapalan, Farmer, D. (2003). Climate, soil and vegetation controls upon the variability of water balance in temperate and semiarid landscapes: downward approach to water balance analysis. Water Resources Research 39(2), p 1035.

Author

D.E. Beaudette