dist.along.grad {sharpshootR}	R Documentation

Compute Euclidean distance along a gradient.

Description

This function computes Euclidean distance along points aligned to a given gradient (e.g. elevation).

Usage


dist.along.grad(coords, var, grad.scaled.min, grad.scaled.max)

Arguments

coords
a matrix of x and y coordinates in some projected coordinate system

var
a vector of the same length as coords, describing the gradient of interest

grad.scaled.min
min value of rescaled gradient values

grad.scaled.max
max value of rescaled gradient values

Details

This function is primarily intended for use within plotTransect.

Value

A data.frame object:

scaled.grad

scaled gradient values

scaled.distance

cumulative distance, scaled to the interval of 0.5, nrow(coords) + 0.5

distance

cumulative distance computed along gradient, e.g. transect distance

variable

sorted gradient values

x

x coordinates, ordered by gradient values

y

y coordinate, ordered by gradient values

grad.order

a vector index describing the sort order defined by gradient values

Note

This function is very much a work in progress, ideas welcome.

Author(s)

D.E. Beaudette

See Also

plotTransect

[Package sharpshootR version 0.8-4 Index]