A very simple estimation of soil moisture state based on volumetric water content

This is a very simple classification of volumetric water content (VWC) into 5 "moisture states", based on an interpretation of water retention thresholds. Classification is performed using VWC at satiation, field capacity (typically 1/3 bar suction), permanent wilting point (typically 15 bar suction), and water surplus in mm. The inputs to this function are closely aligned with the assumptions and output from hydromad::hydromad(sma = 'bucket', ...).

Soil moisture classification rules are as follows:

VWC <= pwp: "very dry"

VWC > pwp and <= (mid-point between fc and pwp): "dry"

VWC > (mid-point between fc and pwp) and <= fc: "moist"

VWC > fc: "very moist"

U (surplus) > 2mm: "saturated / runoff"

estimateSoilMoistureState(VWC, U, sat, fc, pwp)

Arguments

VWC
vector of volumetric water content (VWC), range is 0-1

U
vector of surplus water (mm)

sat
satiation water content, range is 0-1

fc
field capacity water content, range is 0-1

pwp
permanent wilting point water content, range is 0-1

Value

vector of moisture states (ordered factor)

Author

D.E. Beaudette

Examples

# "very moist"
estimateSoilMoistureState(VWC = 0.3, U = 0, sat = 0.35, fc = 0.25, pwp = 0.15)

#> [1] very moist
#> Levels: very dry < dry < moist < very moist < saturated / runoff

# "very dry"
estimateSoilMoistureState(VWC = 0.15, U = 0, sat = 0.35, fc = 0.25, pwp = 0.15)

#> [1] very dry
#> Levels: very dry < dry < moist < very moist < saturated / runoff