Geomorphic Position Probability via SDA

Hillslope position probability estimates from the SDA query service (SSURGO)

hillslopeProbability(s, replaceNA=TRUE)
surfaceShapeProbability(s, replaceNA=TRUE)
geomPosHillProbability(s, replaceNA=TRUE)
geomPosMountainProbability(s, replaceNA=TRUE)

Arguments

s
a character vector of soil series names, automatically normalized to upper case

replaceNA
boolean: should missing classes be converted to probabilities of 0?

Details

These functions send a query to the SDA webservice. Further information on the SDA webservice and query examples can be found at http://sdmdataaccess.nrcs.usda.gov/QueryHelp.aspx

Value

A data.frame object with rows representing soil series, and columns representing probability estimates of that series occuring at specificed geomorphic positions or associated with a surface shape.

Note

Probability values are computed from SSURGO data.

Author

D.E. Beaudette

Examples

# \donttest{

if(requireNamespace("curl") &
   curl::has_internet() &
   require(soilDB)) {
  
  # soil series of interest
s <- c('amador', 'peters', 'pentz', 'inks', 'auburn', 'dunstone', 'argonaut')

# generate hillslope probability table
hillslopeProbability(s)

# generate surface 2D shape probability table
surfaceShapeProbability(s)
  
}

#> single result set, returning a data.frame

#> single result set, returning a data.frame

#>   compname Convex/Convex Linear/Convex Convex/Linear Concave/Convex
#> 1   AMADOR          0.50          0.10          0.00              0
#> 2 ARGONAUT          0.03          0.00          0.00              0
#> 3   AUBURN          0.03          0.08          0.02              0
#> 4 DUNSTONE          0.06          0.00          0.06              0
#> 5     INKS          0.30          0.20          0.00              0
#> 6    PENTZ          0.70          0.02          0.01              0
#> 7   PETERS          0.00          0.00          0.00              0
#>   Linear/Linear Concave/Linear Convex/Concave Linear/Concave Concave/Concave
#> 1          0.30           0.07           0.00           0.00            0.03
#> 2          0.78           0.00           0.01           0.05            0.12
#> 3          0.06           0.00           0.68           0.06            0.07
#> 4          0.17           0.00           0.33           0.11            0.28
#> 5          0.10           0.00           0.25           0.00            0.15
#> 6          0.16           0.00           0.03           0.05            0.02
#> 7          0.29           0.04           0.16           0.45            0.07

# }