hillslope.probability {sharpshootR}	R Documentation

Hillslope Probability via SDA

Description

Hillslope position probability estimates from the SDA query service (SSURGO)

Usage

hillslope.probability(s)

Arguments

s
a character vector of soil series names, in lower-case

Details

This function constructs and executes a query that is sent to the http://sdmdataaccess.nrcs.usda.gov webservice. Further information on the SDA webservice and query examples can be found at http://sdmdataaccess.nrcs.usda.gov/QueryHelp.aspx

Value

A data.frame object with rows representing soil series, and columns representing probability estimates of that series occuring along the standard 2D hillslope positions.

Note

Probability values are computed from SSURGO data.

Author(s)

D.E. Beaudette

Examples


## Not run: 
##D library(soilDB)
##D 
##D # soil series of interest
##D s <- c('auburn', 'pardee', 'amador', 'pentz')
##D 
##D # generate hillslope probability table
##D hillslope.probability(s)
## End(Not run)


[Package sharpshootR version 0.4-1 Index]