joinAdjacency {sharpshootR}	R Documentation

Join Document Adjacency

Description

Convert a set of line segment "join decisions" into a weighted adjacency matrix describing which map unit symbols touch.

Usage


joinAdjacency(x, vars = c("l_musym", "r_musym"))

Arguments

x
a SpatialLinesDataFrame object, with 1 line segment per feature (e.g. simple features)

vars
a vector of two characters naming columns containing "left", and "right" map unit symbols

Value

A weighted adjacency matrix is returned, suitable for plotting directly with plotSoilRelationGraph.

Author(s)

D.E. Beaudette

See Also

plotSoilRelationGraph

[Package sharpshootR version 0.8-4 Index]