Monthly Water Balances

Perform a monthly water balance by "leaky bucket" model, inspired by code from bucket.sim of hydromad package, as defined in Bai et al., (2009) (model "SMA_S1"). The plant available water-holding storage (soil thickness * awc) is used as the "bucket capacity". All water in excess of this capacity is lumped into a single "surplus" term.

monthlyWB(
  AWC,
  PPT,
  PET,
  S_init = 1,
  starting_month = 1,
  rep = 1,
  keep_last = FALSE
)

monthlyWB_summary(w)

Arguments

AWC

available water-holding capacity (mm), typically thickness (mm) * awc (fraction)

PPT

time-series of monthly PPT (mm), calendar year ordering

PET

time-series of monthly PET (mm), calendar year ordering

S_init

initial fraction of AWC filled with water (values 0-1)

starting_month

starting month index, 1=January, 9=September

rep

number of cycles to run water balance

keep_last

keep only the last iteration of the water balance

w

used for for monthlyWB_summary(): a data.frame, such as result of monthlyWB();

Value

a data.frame with the following elements:

PPT: monthly PPT (mm)

PET: monthly PET (mm)

U: monthly surplus (mm)

S: monthly soil moisture storage (mm)

ET: monthly AET (mm)

D: monthly deficit (mm)

month: month number

mo: month label

monthlyWB_summary(): a data.frame containing:

cumulative (dry, moist, wet) days

consecutive (dry_con, moist_con, wet_con) days

total deficit (total_deficit) in mm

total surplus (total_surplus) in mm

total actual evapotranspiration (total_AET) in mm

annual actual evapotranspiration to potential evapotranspiration ratio (annual_AET_PET_ratio)

Details

See the monthly water balance tutorial for further examples and discussion.

A number of important assumptions are made by this style of water balance modeling:

the concept of field capacity is built into the specified bucket size

the influence of aquitards or local terrain cannot be integrated into this model

interception is not used in this model

References

Arkley R, Ulrich R. 1962. The use of calculated actual and potential evapotranspiration for estimating potential plant growth. Hilgardia 32(10):443-469.

Bai, Y., T. Wagener, P. Reed (2009). A top-down framework for watershed model evaluation and selection under uncertainty. Environmental Modelling and Software 24(8), pp. 901-916.

Farmer, D., M. Sivapalan, Farmer, D. (2003). Climate, soil and vegetation controls upon the variability of water balance in temperate and semiarid landscapes: downward approach to water balance analysis. Water Resources Research 39(2), p 1035.