Plot soil profiles below a dendrogram

Plot soil profiles below a dendrogram

plotProfileDendrogram(x, clust, scaling.factor=0.01, width=0.1, y.offset=0.1, 
dend.y.scale= max(clust$height * 2, na.rm=TRUE) , 
dend.color=par('fg'), dend.width=1, debug=FALSE, ...)

Arguments

x
a SoilProfileCollection object

clust
a hierachical clustering object generated by hclust(), cluster::agnes(), or cluster::diana()

scaling.factor
vertical scaling of the profile heights (may have to tinker with this)

width
scaling of profile widths

y.offset
vertical offset for top of profiles

dend.y.scale
extent of y-axis (may have to tinker with this)

dend.color
dendrogram line color

dend.width
dendrogram line width

debug
optionally print debugging data

...
additional arguments to plotSPC

Details

This function places soil profile sketches below a dendrogram.

Note

You may have to tinker with some of the arguments to get optimal arrangement and scaling of soil profiles.

Author

D.E. Beaudette

See also