Visualize Soil Relationships via Chord Diagram

Visualize Soil Relationships via Chord Diagram

plotSoilRelationChordGraph(
  m,
  s,
  mult = 2,
  base.color = "grey",
  highlight.colors = c("RoyalBlue", "DarkOrange", "DarkGreen"),
  add.legend = TRUE,
  ...
)

Arguments

m

an adjacency matrix, no NA allowed

s

soil of interest, must exist in the column or row names of m

mult

multiplier used to re-scale data in m associated with s

base.color

color for all soils other than s and 1st and 2nd most commonly co-occurring soils

highlight.colors

vector of 3 colors: soil of interest, 1st most common, 2nd most common

add.legend

logical, add a legend

...

additional arguments passed to circlize::chordDiagramFromMatrix

Value

nothing, function is called to generate graphical output

Details

This function is experimental. Documentation pending. See http://jokergoo.github.io/circlize/ for ideas.

Author

D.E. Beaudette