Line / Area Visualization for Monthly Water Balance

Pending.

plotWB_lines(
  WB,
  cols = c("#759CC9", "#EB6D6E", "#7FC47D"),
  line.col = "black",
  line.lty = c(1, 2, 3),
  interpolator = c("spline", "linear"),
  spline.method = c("natural", "periodic"),
  month.cex = 1,
  legend.cex = 0.9
)

Arguments

WB

output from monthlyWB()

cols

vector of three colors used for area under PPT, PET, and AET curves

line.col

single color used for PPT, PET, and AET lines

line.lty

vector of three line styles used for PPT, PET, AET curves

interpolator

spline or linear interpolation of monthly values, use of spline may lead to minor smoothing artifacts in shaded areas

spline.method

when interpolator = 'spline', argument passed to splinefun(..., method = spline.method)

month.cex

scaling factor for month labels

legend.cex

scaling factor for legend

Value

nothing, function is called to generate graphical output

Author

J.M. Skovlin and D.E. Beaudette

Examples

if(requireNamespace('hydromad')) {

## A shallow / droughty soil near Sonora CA 
# 100mm (4") AWC
AWC <- 100
PPT <- c(171, 151, 138, 71, 36, 7, 1, 2, 11, 48, 102, 145)
PET <- c(15.17, 18.26, 30.57, 42.95, 75.37, 108.05, 139.74, 128.9, 93.99, 59.84, 26.95, 14.2)

# calendar-year
# three year warm-up
x.wb <- monthlyWB(AWC, PPT, PET, S_init = 0, starting_month = 1, rep = 3, keep_last = TRUE)
 
# plot
plotWB_lines(x.wb)

}
#> Loading required namespace: hydromad