Get and prepare basic soil hydraulic parameters from SSURGO via SDA

Get and prepare basic soil hydraulic parameters from SSURGO via SDA

prepare_SSURGO_hydro_data(cokeys, max.depth)

Arguments

cokeys

vector of component keys (cokey) in current SSURGO snapshot

max.depth

target depth of aggregation (cm), corrected later by real soil depth as reported by slab()

Value

a list containing:

SPC: SoilProfileCollection

agg: aggregate representation of hydraulic parameters, by cokey

The following soil hydraulic properties are included:

variable	description

cokey	component key

hzname	horizon name

hz_top	horizon top depth (cm)

hz_bottom	horizon bottom depth (cm)

thick	horizon thickness (cm)

sat	VWC at saturation (cm/cm)

fc	VWC at field capacity defined by 1/3rd bar tension (cm/cm)

fc_tenthbar	VWC at field capacity defined by 1/3rd bar tension (cm/cm)

pwp	VWC at permanent wilting point or 15 bar tension (cm/cm)

awc	total sand content (<2mm fraction, mass %)

sand	total silt content (<2mm fraction, mass %)

silt	total clay content (<2mm fraction, mass %)

clay	total sand content (<2mm fraction, mass %)

dbthirdbar	bulk density at 1/3 bar tension (g/cm^3)

dbovendry	bulk density oven-dry basis (g/cm^3)

ksat	Ksat (um/second)

soil_fraction	volume fraction of soil (1 - coarse fragment volume fraction)

Details

Weighted mean soil hydraulic parameters are returned over the interval of 0-max.depth, calculated by aqp::slab().

Author

D.E. Beaudette