Sample a Polygon at Fixed Density

Generate sampling points within a SpatialPolygon object, according to a specified sampling density.

sample.by.poly(p, n.pts.per.ac=1, min.samples=5,
sampling.type='regular', iterations=10, p4s=NULL)

Arguments

p
a Polygon object, with coordinates in a projected CRS with units of meters

n.pts.per.ac
requested sampling density in points per acre (results will be close)

min.samples
minimum requested number of samples per polygon

sampling.type
sampling type, see spsample

iterations
number of tries that spsample will attempt

p4s
a qualified proj4string that will be assigned to sampling points

Details

This function is typically accessed via some kind of helper function such as constantDensitySampling.

Value

A SpatialPoints object.

Note

This function expects that the Polygon object has coordinates associated with a projected CRS-- e.g. units of meters. Invalid geometries may cause errors or yield incorrect sample sizes.

See also