Simple interface to the hydromad "leaky bucket" soil moisture model

Simple interface to the hydromad "leaky bucket" soil moisture model.

simpleWB(
  PPT,
  PET,
  D,
  thickness,
  sat,
  fc,
  pwp,
  S_0 = 0.5,
  a.ss = 0.05,
  M = 0,
  etmult = 1
)

Arguments

PPT

precipitation series (mm)

PET

potential ET series (mm)

D

dates

thickness

soil thickness (cm)

sat

volumetric water content at saturation (satiated water content)

fc

volumetric water content at field capacity (typically 1/3 bar suction)

pwp

volumetric water content at permanent wilting point (typically 15 bar suction)

S_0

initial soil moisture as a fraction of total water storage (mm)

a.ss

recession coefficients for subsurface flow from saturated zone, should be > 0

M

fraction of area covered by deep-rooted vegetation

etmult

multiplier for PET

Value

a data.frame

Details

Adjustments for coarse fragments should be made by reducing thickness.

References

Farmer, D., M. Sivapalan, Farmer, D. (2003). Climate, soil and vegetation controls upon the variability of water balance in temperate and semiarid landscapes: downward approach to water balance analysis. Water Resources Research 39(2), p 1035.

Bai, Y., T. Wagener, P. Reed (2009). A top-down framework for watershed model evaluation and selection under uncertainty. Environmental Modelling and Software 24(8), pp. 901-916.