LL2PLSS

Uses latitude and longitude coordinates to return the PLSS section geometry from the BLM PLSS web service.

LL2PLSS(x, y, returnlevel = c("I", "S"))

Arguments

x

longitude coordinates (WGS84)

y

latitude coordinates (WGS84)

returnlevel

'S' for "Section" or 'I' for "Intersection" (subsections)

Value

sf object with geometry and PLSS definition.

Details

This function accepts geographic coordinates and returns the PLSS fabric geometry to the quarter-quarter section. returnlevel defaults to 'I' which returns smallest intersected sectional aliquot geometry, 'S' will return the section geometry of the coordinates. See https://gis.blm.gov/arcgis/rest/services/Cadastral/BLM_Natl_PLSS_CadNSDI/MapServer for details.

Note

This function requires the following packages: httr, jsonlite, and sp.

See also

Author

D.E. Beaudette, Jay Skovlin, A.G. Brown