aggregateColorPlot {sharpshootR}	R Documentation

Plot aggregate soil color data

Description

Generate a plot from summaries generated by aqp::aggregateColor().

Usage


aggregateColorPlot(x, print.label=TRUE, label.font = 1, 
label.cex = 0.65, buffer.pct = 0.02, print.n.hz=FALSE, 
rect.border='black', horizontal.borders=FALSE, ...)

Arguments

x
a list, results from aqp::aggregateColor()

print.label
print Munsell color labels inside of rectangles, when they fit

label.font
font specification for color labels

label.cex
font size for color lables

buffer.pct
extra space between labels and color rectangles

print.n.hz
optionally print the number of horizons

rect.border
color for rectangle border

horizontal.borders
optionally add horizontal borders between bands of color

...
additional arguments passed to plot

Details

See examples.

Author(s)

D.E. Beaudette

Examples

## Not run: 
##D data(loafercreek, package = 'soilDB')
##D 
##D # generalize horizon names using REGEX rules
##D n <- c('Oi', 'A', 'BA','Bt1','Bt2','Bt3','Cr','R')
##D p <- c('O', '^A$|Ad|Ap|AB','BA$|Bw', 
##D 'Bt1$|^B$','^Bt$|^Bt2$','^Bt3|^Bt4|CBt$|BCt$|2Bt|2CB$|^C$','Cr','R')
##D loafercreek$genhz <- generalize.hz(loafercreek$hzname, n, p)
##D 
##D # remove non-matching generalized horizon names
##D loafercreek$genhz[loafercreek$genhz == 'not-used'] <- NA
##D loafercreek$genhz <- factor(loafercreek$genhz)
##D 
##D # aggregate color data, this function is from the `aqp` package
##D a <- aggregateColor(loafercreek, 'genhz')
##D 
##D # plot
##D par(mar=c(1,4,4,1))
##D aggregateColorPlot(a, print.n.hz = TRUE)
## End(Not run)

[Package sharpshootR version 0.8-4 Index]