Diagnostic Property Plot (base graphics)

Generate a graphical description of the presence/absence of soil diagnostic properties.

diagnosticPropertyPlot(
  f,
  v,
  k,
  grid.label = "upedonid",
  dend.label = "upedonid",
  sort.vars = TRUE
)

Arguments

f

SoilProfileCollection object

v

character vector of site-level attribute names of logical type

k

an integer, number of groups to highlight

grid.label

the name of a site-level attribute (usually unique) annotating the y-axis of the grid

dend.label

the name of a site-level attribute (usually unique) annotating dendrogram terminal leaves

sort.vars

sort variables according to natural clustering (TRUE), or use supplied ordering in v

Value

a list is silently returned by this function, containing:

rd

a data.frame containing IDs and grouping code

profile.order

a vector containing the order of soil profiles (row-order in figure), according to diagnostic property values

var.order

a vector containing the order of variables (column-order in figure), according to their distribution among profiles

Details

This function attempts to display several pieces of information within a single figure. First, soil profiles are sorted according to the presence/absence of diagnostic features named in v. Second, these diagnostic features are sorted according to their distribution among soil profiles. Third, a binary grid is established with row-ordering of profiles based on step 1 and column-ordering based on step 2. Blue cells represent the presence of a diagnostic feature. Soils with similar diagnostic features should 'clump' together. See examples below.

See also

Author

D.E. Beaudette and J.M. Skovlin