A very simple estimation of soil moisture state based on volumetric water content

This is a very simple classification of volumetric water content (VWC) into 5 "moisture states", based on an interpretation of water retention thresholds. Classification is performed using VWC at satiation, field capacity (typically 1/3 bar suction), permanent wilting point (typically 15 bar suction), and water surplus in mm. The inputs to this function are closely aligned with the assumptions and output from hydromad::hydromad(sma = 'bucket', ...).

Soil moisture classification rules are as follows:

VWC <= pwp: "very dry"

VWC > pwp AND <= (mid-point between fc and pwp): "dry"

VWC > (mid-point between fc and pwp) AND <= fc: "moist"

VWC > fc: "very moist"

VWC > fc AND U (surplus) > 4mm: "wet"

estimateSoilMoistureState(
  VWC,
  U,
  sat,
  fc,
  pwp,
  style = c("default", "newhall")
)

Arguments

VWC

vector of volumetric water content (VWC), range is 0-1

U

vector of surplus water (mm)

sat

satiation water content, range is 0-1

fc

field capacity water content, range is 0-1

pwp

permanent wilting point water content, range is 0-1

style

VWC classification style

Value

vector of moisture states (ordered factor)

Author

D.E. Beaudette

Examples

# "very moist"
estimateSoilMoistureState(VWC = 0.3, U = 0, sat = 0.35, fc = 0.25, pwp = 0.15)
#> [1] very moist
#> Levels: very dry < dry < moist < very moist < wet
estimateSoilMoistureState(VWC = 0.3, U = 2, sat = 0.35, fc = 0.25, pwp = 0.15)
#> [1] very moist
#> Levels: very dry < dry < moist < very moist < wet

"wet"
#> [1] "wet"
estimateSoilMoistureState(VWC = 0.3, U = 5, sat = 0.35, fc = 0.25, pwp = 0.15)
#> [1] wet
#> Levels: very dry < dry < moist < very moist < wet

# "very dry"
estimateSoilMoistureState(VWC = 0.15, U = 0, sat = 0.35, fc = 0.25, pwp = 0.15)
#> [1] very dry
#> Levels: very dry < dry < moist < very moist < wet

# "dry" 
estimateSoilMoistureState(VWC = 0.18, U = 0, sat = 0.35, fc = 0.25, pwp = 0.15)
#> [1] dry
#> Levels: very dry < dry < moist < very moist < wet