formatPLSS

Format PLSS information into a coded format that can be digested by PLSS web service.

formatPLSS(p, type = "SN")

Arguments

p

data.frame with chunks of PLSS coordinates

type

an option to format protracted blocks 'PB', unprotracted blocks 'UP', or standard section number 'SN' (default).

Value

A vector of PLSS codes.

Details

This function is typically accessed as a helper function to prepare data for use within PLSS2LL function.

Note

This function expects that the Polygon object has coordinates associated with a projected CRS-- e.g. units of meters.

This function requires the following packages: stringi.

See also

Author

D.E. Beaudette, Jay Skovlin, A.G. Brown

Examples

# create some data
d <- data.frame(
  id = 1:3,
  qq = c('SW', 'SW', 'SE'),
  q = c('NE', 'NW', 'SE'),
  s = c(17, 32, 30),
  t = c('T36N', 'T35N', 'T35N'),
  r = c('R29W', 'R28W', 'R28W'),
  type = 'SN',
  m = 'MT20',
  stringsAsFactors = FALSE
)
# add column names

names(d) <- c('id', 'qq', 'q', 's', 't', 'r', 'type', 'm')
# generate formatted PLSS codes
formatPLSS(d, type='SN')
#> [1] "MT200360N0290W0SN170ASWNE" "MT200350N0280W0SN320ASWNW"
#> [3] "MT200350N0280W0SN300ASESE"