hillslope.probability {sharpshootR}	R Documentation

Hillslope Probability via SDA

Description

Hillslope position probability estimates from the SDA query service (SSURGO)

Usage

hillslope.probability(s)

Arguments

s
a character vector of soil series names, in lower-case

Details

This function constructs and executes a query that is sent to the http://sdmdataaccess.nrcs.usda.gov webservice. Further information on the SDA webservice and query examples can be found at http://sdmdataaccess.nrcs.usda.gov/QueryHelp.aspx

Value

A data.frame object with rows representing soil series, and columns representing probability estimates of that series occuring along the standard 2D hillslope positions.

Note

Probability values are computed from SSURGO data.

Author(s)

D.E. Beaudette

Examples


require(soilDB)
require(reshape)

# soil series of interest
s <- c('auburn', 'pardee', 'amador', 'pentz')

# generate hillslope probability table
hillslope.probability(s)

## sending SOAP request...
## processing results...
## 
  |                                                                       
  |                                                                 |   0%
  |                                                                       
  |=====                                                            |   7%
  |                                                                       
  |=========                                                        |  14%
  |                                                                       
  |==============                                                   |  21%
  |                                                                       
  |===================                                              |  29%
  |                                                                       
  |=======================                                          |  36%
  |                                                                       
  |============================                                     |  43%
  |                                                                       
  |================================                                 |  50%
  |                                                                       
  |=====================================                            |  57%
  |                                                                       
  |==========================================                       |  64%
  |                                                                       
  |==============================================                   |  71%
  |                                                                       
  |===================================================              |  79%
  |                                                                       
  |========================================================         |  86%
  |                                                                       
  |============================================================     |  93%
  |                                                                       
  |=================================================================| 100%

##   compname Toeslope Footslope Backslope Shoulder Summit
## 1   amador       NA      0.03      0.45       NA   0.21
## 2   auburn       NA      0.03      0.70     0.04   0.17
## 3   pardee     0.18        NA      0.27     0.09   0.36
## 4    pentz     0.02        NA      0.58       NA   0.39



[Package sharpshootR version 0.4 Index]