Vizualize Soil Relationships via Chord Diagram.

Plot a soil relationship diagram using a chord diagram.

plotSoilRelationChordGraph(m, s, mult = 2, base.color = "grey",
highlight.colors = c("RoyalBlue", "DarkOrange", "DarkGreen"),
add.legend = TRUE, ...)

Arguments

m
an adjacency matrix, no NA allowed

s
soil of interest, must exist in the column or row names of m

mult
multiplier used to re-scale data in m associated with s

base.color
color for all soils other than s and 1st and 2nd most commonly co-occuring

highlight.colors
vector of 3 colors: soil of interest, 1st most common, 2nd most common

add.legend
logical, add a legend

…
additional arguments passed to circlize::chordDiagramFromMatrix

Details

This function is experimental. Documentation pending. See http://jokergoo.github.io/circlize/ for ideas.

References

https://github.com/jokergoo/circlize

See also

Examples

# NOT RUN {
data(amador)
m <- component.adj.matrix(amador)
plotSoilRelationChordGraph(m, 'amador')
# }