sample.by.poly {sharpshootR}	R Documentation

Sample by Polygon

Description

Sample by polygon, according to a specified sampling density

Usage

sample.by.poly(p, n.pts.per.ac = 1, min.samples = 5, 
sampling.type = "hexagonal", iterations = 10, ...)

Arguments

p
a Polygon object, with coordinates from a projected CRS

n.pts.per.ac
requested sampling density in points per acre (results will be close)

min.samples
minimum requested samples / polygon

sampling.type
sampling type, see spsample

iterations
number of tries that spsample will attempt

...
additional arguments to spsample

Details

This function is typically used within the context of a loop or similar function that iterates over a list of Polygons associated with a SpatialPolygons object. See examples.

Value

A SpatialPoints object.

Note

This function expects that the Polygon object has coordinates associated with a projected CRS– e.g. units of meters.

Author(s)

D.E. Beaudette

[Package sharpshootR version 0.7 Index]