Annual Climate Summaries for Soil Series Data

Annual climate summaries for soil series, based on latticeExtra::segplot, based on 5th, 25th, 50th, 75th, and 95th percentiles. Input data should be from soilDB::fetchOSD.

vizAnnualClimate(climate.data, IQR.cex = 1, s = NULL, s.col = "firebrick", ...)

Arguments

climate.data

Annual climate summaries, as returned from soilDB::fetchOSD(..., extended = TRUE), should not contain NA values

IQR.cex

scaling factor for bar representing interquartile range

s

a soil series name, e.g. "LUCY", to highlight

s.col

color for highlighted soil series

...

further arguments passed to latticeExtra::segplot

Value

A list with the following elements:

fig: lattice object (the figure)

clust: clustering object returned by cluster::diana

Details

This function was designed for use with soilDB::fetchOSD(). It might be possible to use with other sources of data but your mileage may vary. See the Soil Series Query Functions tutorial for more information.

See also

Author

D.E. Beaudette